moved the welcome page to index.tsx and removed signin/signup from homepage
